Move SendGrid GetMailMessageEvent dispatching to SendGridApiFacade
Moving the dispatcher injection and event dispatching functionality to
the SendGridApiFacade allows for better access to the API interaction
process.

Dispatching the event immediately after the the SendGridApiMessage returns the
message will have the same effect as the previous approach within the
getMessage method.
