+++ app/bundles/EmailBundle/Swiftmailer/SendGrid/SendGridApiMessage.php
@@ -11,13 +11,11 @@
 
 namespace Mautic\EmailBundle\Swiftmailer\SendGrid;
 
-use Mautic\EmailBundle\Swiftmailer\SendGrid\Event\GetMailMessageEvent;
 use Mautic\EmailBundle\Swiftmailer\SendGrid\Mail\SendGridMailAttachment;
 use Mautic\EmailBundle\Swiftmailer\SendGrid\Mail\SendGridMailBase;
 use Mautic\EmailBundle\Swiftmailer\SendGrid\Mail\SendGridMailMetadata;
 use Mautic\EmailBundle\Swiftmailer\SendGrid\Mail\SendGridMailPersonalization;
 use SendGrid\Mail;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 class SendGridApiMessage
 {
@@ -41,23 +39,16 @@ class SendGridApiMessage
      */
     private $sendGridMailAttachment;
 
-    /**
-     * @var EventDispatcherInterface
-     */
-    private $dispatcher;
-
     public function __construct(
         SendGridMailBase $sendGridMailBase,
         SendGridMailPersonalization $sendGridMailPersonalization,
         SendGridMailMetadata $sendGridMailMetadata,
-        SendGridMailAttachment $sendGridMailAttachment,
-        EventDispatcherInterface $dispatcher
+        SendGridMailAttachment $sendGridMailAttachment
     ) {
         $this->sendGridMailBase            = $sendGridMailBase;
         $this->sendGridMailPersonalization = $sendGridMailPersonalization;
         $this->sendGridMailMetadata        = $sendGridMailMetadata;
         $this->sendGridMailAttachment      = $sendGridMailAttachment;
-        $this->dispatcher                  = $dispatcher;
     }
 
     /**
@@ -73,21 +64,6 @@ class SendGridApiMessage
         $this->sendGridMailMetadata->addMetadataToMail($mail, $message);
         $this->sendGridMailAttachment->addAttachmentsToMail($mail, $message);
 
-        $this->dispatchGetMailMessageEvent($mail, $message);
-
         return $mail;
-    }
-
-    /**
-     * Dispatch GET_MAIL_MESSAGE event.
-     *
-     * @param Mail                $mail
-     * @param \Swift_Mime_Message $message
-     */
-    private function dispatchGetMailMessageEvent(Mail $mail, \Swift_Mime_Message $message)
-    {
-        $event = new GetMailMessageEvent($mail, $message);
-
-        $this->dispatcher->dispatch(SendGridMailEvents::GET_MAIL_MESSAGE, $event);
     }
 }
