+++ app/bundles/EmailBundle/Swiftmailer/SendGrid/SendGridMailEvents.php
@@ -15,9 +15,9 @@ final class SendGridMailEvents
 {
     /**
      * The mautic.email.swiftmailer.sendgrid_api.get_mail_message event is
-     * dispatched by the mautic.transport.sendgrid_api.message service
-     * as the last step before returning a SendGrid\Mail instance in the
-     * getMessage method.
+     * dispatched by the mautic.transport.sendgrid_api.facade service
+     * immediately after obtaining a SendGrid\Mail instance from the
+     * mautic.transport.sendgrid_api.message service.
      *
      * The event listener receives an instance of
      * \Mautic\EmailBundle\Swiftmailer\SendGrid\Event\GetMailMessageEvent.
